Infinite World Architecture - 06 - 02 - 2026
Added:

+ 64-bit chunk coordinates using Vector3D<long> for virtually unlimited world size
+ Architecture documentation for coordinate system (internal long/int, rendering float)
+ TODO comments for future optimizations (block data, visibility culling, face culling, meshing)
+ Commented example method LocalToWorldRelative() for camera-relative rendering
+ Extended XML documentation with architecture notes

Changes:

~ ChunkPosition from Vector3D<int> to Vector3D<long>
~ WorldToChunkPosition() changed to public static (was private)
~ WorldToLocal() changed to public static (was private)
~ Coordinate conversion methods now document internal precision vs rendering precision
~ Updated remarks to explain separation of chunk data and rendering logic

Fixed:

* Typo corrected: WorldTolocal() → WorldToLocal()

Removed:

- 32-bit chunk coordinate limitation (±2.1 billion blocks per axis)

+++ Create your Adventure/Program.cs
@@ -179,7 +179,7 @@ namespace Create_your_Adventure
             Logger.Info("[OPENGL] Vertex Array Object created");
 
             // -------- Instancing Setup --------
-            chunk = new Chunk(new Vector3D<int>(0, 0, 0));
+            chunk = new Chunk(new Vector3D<long>(0, 0, 0));
 
             // -------- Instance VBO --------
             instanceVBO = gl.GenBuffer();
